Add files via upload
This Project File contains Code for project done for udacity nanodegree.
This is a web App that allows Users to draw "Pixel Art" on a canvas(background) customizable to terms of the user! 
The aim of the "Pixel Art Maker" project aims at building amazing colorful Art from a Selection of grids where a user can select Grid size (height and width numbers) and color(s) into different cells of the grid. The page is responsive and can work perfectly well on any device.